Fixing issue #27. Now all folders for views, templates and styles are always treated properly following the help, which suggest that these names should always follow the naming conventions, re-using controller's name - which is plural.
All generators and destroyers was fixed. Destroyers had also another
issues related to the `--rf` option. This issue is now resolved as well.

The `controller.coffee` was also fixed to properly locate it's related
view while rendering views automatically through the generated action
stub - that is used when the user didn't define the action itself.
the action